feat[163]: 주최자 메인 페이지 구현 중
