+++ src/components/Main/MainPage.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import MainPageDesktop from './desktop/MainPageDesktop';
 import Desktop from '../MediaQuery/Desktop';
 import Mobile from '../MediaQuery/Mobile';
+import MainPageDesktop from './desktop/MainPageDesktop';
 import MainPageMobile from './Mobile/MainPageMobile';
 
 function MainPage() {
